Split up results component/route in summary+full
